Separaciónde responsabilidades
Se reparten funciones entre varios archivos (Se podrían sacar varios HandleEvent de ser necesario)
